Fix bug in config loading
