+++ astro/cli/astro/cmd/config.go
@@ -17,10 +17,9 @@
 package cmd
 
 import (
-	"errors"
+	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/uber/astro/astro"
 	"github.com/uber/astro/astro/utils"
 )
 
@@ -33,21 +32,10 @@ var configFileSearchPaths = []string{
 	"terraform/astro.yml",
 }
 
-var errCannotFindConfig = errors.New("unable to find config file")
-
-// loadConfig tries to load astro configuration, either from a user specified
-// path or from one of the search paths. If no astro config can be found, an
-// error is returned.
-func (cli *AstroCLI) loadConfig(configFilePath string) error {
-	config, err := astro.NewConfigFromFile(configFilePath)
-	cli.config = config
-	return err
-}
-
 // configPathFromArgs reads the command line arguments and returns the value of
 // the config option. It returns an empty string if there is no path in the
 // args.
-func configPathFromArgs(args []string) (configFilePath string) {
+func configPathFromArgs(args []string) (configFilePath string, err error) {
 	// this is a special cobra command so that we can parse just the config
 	// flag early in the program lifecycle.
 	findConfig := &cobra.Command{
@@ -71,21 +59,14 @@ func configPathFromArgs(args []string) (configFilePath string) {
 	// Do an early first parse of the config flag before the main command,
 	findConfig.PersistentFlags().StringVar(&configFilePath, "config", "", "config file")
 	if err := findConfig.ParseFlags(finalArgs); err != nil {
-		return ""
+		return "", err
 	}
 
-	return configFilePath
-}
-
-// resolveConfigFilePath returns the path of the project config file. It
-// returns an empty string if no config file is found.
-func (cli *AstroCLI) resolveConfigFilePath() (configFilePath string) {
-	// Try to find the config file
-	if path := firstExistingFilePath(configFileSearchPaths...); path != "" {
-		return path
+	if configFilePath != "" && !utils.FileExists(configFilePath) {
+		return "", fmt.Errorf("%v: file does not exist", configFilePath)
 	}
 
-	return ""
+	return configFilePath, nil
 }
 
 // firstExistingFilePath takes a list of paths and returns the first one
