+++ astro/cli/astro/cmd/fixtures/config-simple/astro.yaml
