+++ astro/cli/astro/cmd/flags_test.go
@@ -24,12 +24,16 @@ import (
 	"github.com/uber/astro/astro/tests"
 )
 
+func TestHelpWorks(t *testing.T) {
+	result := tests.RunTest(t, []string{"--help"}, "fixtures/no-config", tests.VERSION_LATEST)
+	assert.Contains(t, result.Stderr.String(), "A tool for managing multiple Terraform modules")
+	assert.Equal(t, 0, result.ExitCode)
+}
 func TestHelpUserFlags(t *testing.T) {
 	result := tests.RunTest(t, []string{
-		"--config=simple_variables.yaml",
 		"plan",
 		"--help",
-	}, "fixtures/flags", tests.VERSION_LATEST)
+	}, "fixtures/config-simple", tests.VERSION_LATEST)
 	assert.Contains(t, result.Stderr.String(), "User flags:")
 	assert.Contains(t, result.Stderr.String(), "--foo")
 	assert.Contains(t, result.Stderr.String(), "--baz")
@@ -46,13 +50,13 @@ func TestHelpNoUserFlags(t *testing.T) {
 	assert.NotContains(t, result.Stderr.String(), "User flags:")
 }
 
-func TestConfigLoadError(t *testing.T) {
+func TestConfigLoadErrorWhenSpecified(t *testing.T) {
 	result := tests.RunTest(t, []string{
 		"--config=/nonexistent/path/to/config",
 		"plan",
 		"--help",
-	}, "fixtures/flags", tests.VERSION_LATEST)
-	assert.Contains(t, result.Stderr.String(), "no such file or directory")
+	}, "fixtures/config-simple", tests.VERSION_LATEST)
+	assert.Contains(t, result.Stderr.String(), "file does not exist")
 	assert.Equal(t, 1, result.ExitCode)
 }
 
@@ -68,9 +72,8 @@ func TestUnknownFlag(t *testing.T) {
 
 func TestPlanErrorOnMissingValues(t *testing.T) {
 	result := tests.RunTest(t, []string{
-		"--config=simple_variables.yaml",
 		"plan",
-	}, "fixtures/flags", tests.VERSION_LATEST)
+	}, "fixtures/config-simple", tests.VERSION_LATEST)
 	assert.Equal(t, 1, result.ExitCode)
 	assert.Contains(t, result.Stderr.String(), "missing required flags")
 	assert.Contains(t, result.Stderr.String(), "--foo")
